backend/requirements.txt
# Web framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File upload handling
python-multipart==0.0.6

# Image processing
pillow==10.1.0

# Data validation and serialization
pydantic==2.5.1

# Note: ffmpeg is required as system dependency
# Install with: brew install ffmpeg (macOS) or apt install ffmpeg (Ubuntu)